Refactor Muon optimiser location
